refactors code to remove bad imports and create-knative-utils
