+++ frontend/packages/knative-plugin/src/components/overview/__tests__/KnativeResourceOverviewPage.spec.tsx
@@ -4,7 +4,7 @@ import { OverviewItem } from '@console/shared';
 import { LoadingBox } from '@console/internal/components/utils';
 import { ResourceOverviewDetails } from '@console/internal/components/overview/resource-overview-details';
 import { revisionObj } from '@console/dev-console/src/components/topology/__tests__/topology-knative-test-data';
-import { RevisionModel } from '@console/knative-plugin/src/models';
+import { RevisionModel } from '../../../models';
 import { KnativeResourceOverviewPage } from '../KnativeResourceOverviewPage';
 
 describe('KnativeResourceOverviewPage', () => {
