+++ frontend/packages/knative-plugin/src/index.ts
@@ -1,6 +1,4 @@
 export * from './const';
 export * from './models';
-export * from './utils/create-knative-utils';
-export * from './utils/get-knative-icon';
 export * from './types';
 export * from './actions';
