+++ frontend/packages/knative-plugin/src/utils/__tests__/create-knative-utils.spec.ts
@@ -1,53 +1,8 @@
-import { FirehoseResource } from '@console/internal/components/utils';
 import { K8sResourceKind, ImagePullPolicy } from '@console/internal/module/k8s';
-import {
-  getKnativeServiceDepResource,
-  knativeServingResourcesServices,
-  knativeServingResourcesRevision,
-  knativeServingResourcesConfigurations,
-  knativeServingResourcesRoutes,
-} from '../create-knative-utils';
+import { getKnativeServiceDepResource } from '../create-knative-utils';
 import { defaultData } from './knative-serving-data';
 
 describe('Create knative Utils', () => {
-  describe('knative Serving Resources', () => {
-    const SAMPLE_NAMESPACE = 'mynamespace';
-    it('expect knativeServingResource to return service with proper namespace', () => {
-      const serviceServingResource: FirehoseResource[] = knativeServingResourcesServices(
-        SAMPLE_NAMESPACE,
-      );
-      expect(serviceServingResource).toHaveLength(1);
-      expect(serviceServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
-      expect(serviceServingResource[0].prop).toBe('ksservices');
-    });
-
-    it('expect knativeServingResourcesRevision to return revision with proper namespace', () => {
-      const revisionServingResource: FirehoseResource[] = knativeServingResourcesRevision(
-        SAMPLE_NAMESPACE,
-      );
-      expect(revisionServingResource).toHaveLength(1);
-      expect(revisionServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
-      expect(revisionServingResource[0].prop).toBe('revisions');
-    });
-
-    it('expect knativeServingResourcesConfigurations to return configurations with proper namespace', () => {
-      const configServingResource: FirehoseResource[] = knativeServingResourcesConfigurations(
-        SAMPLE_NAMESPACE,
-      );
-      expect(configServingResource).toHaveLength(1);
-      expect(configServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
-      expect(configServingResource[0].prop).toBe('configurations');
-    });
-
-    it('expect knativeServingResourcesRoutes to return routes with proper namespace', () => {
-      const routeServingResource: FirehoseResource[] = knativeServingResourcesRoutes(
-        SAMPLE_NAMESPACE,
-      );
-      expect(routeServingResource).toHaveLength(1);
-      expect(routeServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
-      expect(routeServingResource[0].prop).toBe('ksroutes');
-    });
-  });
   describe('knative Service creation', () => {
     it('expect knativeServingResourcesConfigurations to not have Ports id if unknownTargetPort is empty', () => {
       const knDeploymentResource: K8sResourceKind = getKnativeServiceDepResource(
