+++ frontend/packages/knative-plugin/src/utils/__tests__/get-knative-resources.spec.ts
@@ -1,3 +1,4 @@
+import { FirehoseResource } from '@console/internal/components/utils';
 import { MockResources } from '@console/dev-console/src/components/topology/__tests__/topology-test-data';
 import { MockKnativeResources } from '../../topology/__tests__/topology-knative-test-data';
 import {
@@ -10,6 +11,10 @@ import {
   getEventSourceCamel,
   getEventSourceKafka,
   getEventSourceSinkBinding,
+  knativeServingResourcesServices,
+  knativeServingResourcesRevision,
+  knativeServingResourcesConfigurations,
+  knativeServingResourcesRoutes,
 } from '../get-knative-resources';
 import {
   deploymentData,
@@ -130,6 +135,45 @@ describe('Get knative resources', () => {
     it('expect getEventSourceSinkBinding to return event source as undefined', () => {
       const knEventResource = getEventSourceSinkBinding(deploymentData, MockResources);
       expect(knEventResource).toBeUndefined();
+    });
+  });
+
+  describe('knative Serving Resources', () => {
+    const SAMPLE_NAMESPACE = 'mynamespace';
+    it('expect knativeServingResource to return service with proper namespace', () => {
+      const serviceServingResource: FirehoseResource[] = knativeServingResourcesServices(
+        SAMPLE_NAMESPACE,
+      );
+      expect(serviceServingResource).toHaveLength(1);
+      expect(serviceServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
+      expect(serviceServingResource[0].prop).toBe('ksservices');
+    });
+
+    it('expect knativeServingResourcesRevision to return revision with proper namespace', () => {
+      const revisionServingResource: FirehoseResource[] = knativeServingResourcesRevision(
+        SAMPLE_NAMESPACE,
+      );
+      expect(revisionServingResource).toHaveLength(1);
+      expect(revisionServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
+      expect(revisionServingResource[0].prop).toBe('revisions');
+    });
+
+    it('expect knativeServingResourcesConfigurations to return configurations with proper namespace', () => {
+      const configServingResource: FirehoseResource[] = knativeServingResourcesConfigurations(
+        SAMPLE_NAMESPACE,
+      );
+      expect(configServingResource).toHaveLength(1);
+      expect(configServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
+      expect(configServingResource[0].prop).toBe('configurations');
+    });
+
+    it('expect knativeServingResourcesRoutes to return routes with proper namespace', () => {
+      const routeServingResource: FirehoseResource[] = knativeServingResourcesRoutes(
+        SAMPLE_NAMESPACE,
+      );
+      expect(routeServingResource).toHaveLength(1);
+      expect(routeServingResource[0].namespace).toBe(SAMPLE_NAMESPACE);
+      expect(routeServingResource[0].prop).toBe('ksroutes');
     });
   });
 });
