+++ frontend/packages/knative-plugin/src/utils/create-knative-utils.ts
@@ -1,23 +1,11 @@
 import * as _ from 'lodash';
-import { K8sResourceKind, referenceForModel, ImagePullPolicy } from '@console/internal/module/k8s';
-import { FirehoseResource } from '@console/internal/components/utils';
-import {
-  ServiceModel,
-  RevisionModel,
-  ConfigurationModel,
-  RouteModel,
-  EventSourceCronJobModel,
-  EventSourceContainerModel,
-  EventSourceApiServerModel,
-  EventSourceCamelModel,
-  EventSourceKafkaModel,
-  EventSourceSinkBindingModel,
-} from '@console/knative-plugin';
+import { K8sResourceKind, ImagePullPolicy } from '@console/internal/module/k8s';
 import { getAppLabels, mergeData } from '@console/dev-console/src/utils/resource-label-utils';
 import {
   DeployImageFormData,
   GitImportFormData,
 } from '@console/dev-console/src/components/import/import-types';
+import { ServiceModel } from '../models';
 
 export const getKnativeServiceDepResource = (
   formData: GitImportFormData | DeployImageFormData,
@@ -143,134 +131,4 @@ export const getKnativeServiceDepResource = (
   const knativeDeployResource = mergeData(knativeServiceUpdated || {}, newKnativeDeployResource);
 
   return knativeDeployResource;
-};
-
-export const knativeServingResourcesRevision = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(RevisionModel),
-      namespace,
-      prop: 'revisions',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const knativeServingResourcesConfigurations = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(ConfigurationModel),
-      namespace,
-      prop: 'configurations',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const knativeServingResourcesRoutes = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(RouteModel),
-      namespace,
-      prop: 'ksroutes',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const knativeServingResourcesServices = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(ServiceModel),
-      namespace,
-      prop: 'ksservices',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesCronJob = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceCronJobModel),
-      namespace,
-      prop: 'eventSourceCronjob',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesContainer = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceContainerModel),
-      namespace,
-      prop: 'eventSourceContainers',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesApiServer = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceApiServerModel),
-      namespace,
-      prop: 'eventSourceApiserver',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesCamel = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceCamelModel),
-      namespace,
-      prop: 'eventSourceCamel',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesKafka = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceKafkaModel),
-      namespace,
-      prop: 'eventSourceKafka',
-      optional: true,
-    },
-  ];
-  return knativeResource;
-};
-
-export const eventSourceResourcesSinkBinding = (namespace: string): FirehoseResource[] => {
-  const knativeResource = [
-    {
-      isList: true,
-      kind: referenceForModel(EventSourceSinkBindingModel),
-      namespace,
-      prop: 'eventSourceSinkbinding',
-      optional: true,
-    },
-  ];
-  return knativeResource;
 };
